+++ Assets/Samples/SaveGameSample.cs
@@ -1,21 +1,32 @@
-﻿using UnityEngine;
+﻿using UniSharper.Data.SaveGame.Providers;
+using UnityEngine;
 using UnityEngine.UI;
 
 namespace UniSharper.Data.SaveGame.Samples
 {
+    public enum CipherMode
+    {
+        Ecb,
+        
+        Cbc
+    }
+    
     public class SaveGameSample : MonoBehaviour
     {
-        private const string GameDataName = "SaveGameExample";
-
         private const string Key = "SAnzX2EoRV9Haaqc";
+        
+        [SerializeField]
+        private CipherMode cipherMode = CipherMode.Cbc;
 
         [SerializeField]
-        private InputField inputFiled;
+        private InputField inputField;
 
         [SerializeField]
         private Button loadGameButton;
 
         private ISaveGameManager manager;
+        
+        private string GameDataName => cipherMode == CipherMode.Ecb ? "SaveGameExample" : "SaveGameExampleCbc";
 
         public void OnLoadGameButtonClicked()
         {
@@ -24,18 +35,19 @@ namespace UniSharper.Data.SaveGame.Samples
 
         public void OnSaveGameButtonClicked()
         {
-            if (!string.IsNullOrEmpty(inputFiled.text))
+            if (!string.IsNullOrEmpty(inputField.text))
             {
-                var success = manager.SaveGame(GameDataName, inputFiled.text, true, true);
-                inputFiled.text = string.Empty;
+                var success = manager.SaveGame(GameDataName, inputField.text, true, true);
+                inputField.text = string.Empty;
                 loadGameButton.interactable = success;
             }
         }
 
         private void Awake()
         {
+            ICryptoProvider cryptoProvider = cipherMode == CipherMode.Cbc ? new AesCbcCryptoProvider() : new AesEcbCryptoProvider();
             manager = new SaveGameManager();
-            manager.Initialize();
+            manager.Initialize(null, cryptoProvider);
             loadGameButton.interactable = manager.ExistsSaveData(GameDataName);
         }
 
@@ -45,7 +57,7 @@ namespace UniSharper.Data.SaveGame.Samples
             if (string.IsNullOrEmpty(content))
                 return;
 
-            inputFiled.text = content;
+            inputField.text = content;
         }
     }
 }